patch-list: add inline dropdown for delegate and state one-off changes
Add dropdown for the cell values of the Delegate and State columns for
each individual patch to make one-off changes to patches. The dropdowns
are only viewable if the user has patch edit permissions.

Change the generic_list method to pass the list of states and
maintainers to the patch list view context to populate the dropdown
options. The static patch-list.js file now uses the modularity of the
fetch request and update/error messages handling of rest.js.

TODO: The loading of the patch-list page is very slow now because it
seems that for each call to check if a patch is editable by a user, the
db is accessed. Changes in the backend need to be made so this is done
with only done with only one call to the db.