patch-detail: move patch diff section up and add collapse button
Add new hide/show button to "Patch" header action bar so that the patch
diff can be collapsed for easier viewing of comments and other related
information. Also, the buttons are now left-aligned instead of
right-aligned. The download buttons are also next to the patch subject
title. Before [1] when the Patch section is at the bottom of the page
and after [2] where it is moved to the top and button group next to the
patch subject title is also left-aligned now.

[1] https://imgur.com/svhP18W
[2] https://imgur.com/loFPIpX