patch-detail: add patch relations table
The motivation behind the patch relations table is to be able to provide
information about a given version of a patch in relation to its other
versions. The table indicates whether there exists unaddressed comments
in other versions and can also show what's the most up-to-date version
of a given patch. The information on what is unresolved throughout the
related patches history helps reviewers and submitters keep track of and
quickly locate what needs to be done for a patch. Image [1] for
reference. This involves:

- Add and style patch relations table with a summary row that shows the
  total unaddressed and addressed comments for the patch relation and
  the existence of various tags through their abbreviations (e.g. A/R/T).
- Add collapse button that hides all patches except for the current
  patch (which is highlighted) and the patch relation summary row [2].

If the patch relation contains errors, users can use the "Wrong related
patches?" link to address the issues. For users with patch edit
permissions, the patch relation can be modified (i.e. add/remove patches)
[3]. For users without permission, the ideal behavior is something worth
discussing with various users.

Something that is left TODO is fix the email prompt to maintainers to
include the mailing list for each specific project instead of the
patchwork mailing list. Also, commas and whitespace characters from
the patch subject are cut off in the subject line of the email which
needs to be fixed.

[1] https://i.imgur.com/cwYKiUx.png
[2] https://i.imgur.com/4HGCTYp.png
[3] https://i.imgur.com/Xyy18oL.png