patch-detail: add functionality to add/remove related patches
Currently, patch relations can be modified using the REST API, but the
Patchwork UI provides no tools that allow users to create and edit patch
relations. This patch adds that as part of the patch relations table.
The changes involve the following details:

Add handling of user input to the add/remove related patches as such:
- Parse user input of ids and msgids by commas and whitespace
  - Match parsed ids to patches in the db
    - Record invalid ids to report back to the user
- Update patch relations based on selected option to add/remove
- Add update/error messages to web page as feedback

In the future, an automated system to relate patches together would be
ideal, meaning that this functionality should be considered a fallback
option when the automated system has faults. At the very least, this
manual option serves as the basic function to create and modify patch
relations.